# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProFile API",
    "description": "ProFile Resume & Portfolio Management API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/users/preferences": {
      "get": {
        "operationId": "users_getPreferences",
        "tags": [
          "users"
        ],
        "summary": "Get user preferences (basic)",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get user preferences (basic)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "patch": {
        "operationId": "users_updatePreferences",
        "tags": [
          "users"
        ],
        "summary": "Update user preferences (basic)",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Update user preferences (basic)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/users/preferences/full": {
      "get": {
        "operationId": "users_getFullPreferences",
        "tags": [
          "users"
        ],
        "summary": "Get all user preferences",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get all user preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFullPreferencesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "patch": {
        "operationId": "users_updateFullPreferences",
        "tags": [
          "users"
        ],
        "summary": "Update all user preferences",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Update all user preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFullPreferencesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/v1/users/me/consent-history": {
      "get": {
        "operationId": "user-consent_getConsentHistory",
        "tags": [
          "user-consent"
        ],
        "summary": "Get consent acceptance history",
        "description": "Retrieves all consent records for the authenticated user",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get consent acceptance history"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/users/me/consent-status": {
      "get": {
        "operationId": "user-consent_checkConsentStatus",
        "tags": [
          "user-consent"
        ],
        "summary": "Check consent acceptance status",
        "description": "Returns which documents the user has accepted for the current versions",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Check consent acceptance status"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/users/me/accept-consent": {
      "post": {
        "operationId": "user-consent_acceptConsent",
        "tags": [
          "user-consent"
        ],
        "summary": "Accept Terms of Service or Privacy Policy",
        "description": "Records user acceptance of legal documents with IP and user agent for audit trail. ",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Accept Terms of Service or Privacy Policy"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/v1/mec/institutions": {
      "get": {
        "operationId": "mec-institutions_listInstitutions",
        "tags": [
          "mec-institutions"
        ],
        "summary": "List institutions",
        "parameters": [
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List institutions"
          }
        }
      }
    },
    "/api/v1/mec/institutions/search": {
      "get": {
        "operationId": "mec-institutions_searchInstitutionsByName",
        "tags": [
          "mec-institutions"
        ],
        "summary": "Search institutions",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search institutions"
          }
        }
      }
    },
    "/api/v1/mec/institutions/{codigoIes}": {
      "get": {
        "operationId": "mec-institutions_getInstitutionByCodeWithCourses",
        "tags": [
          "mec-institutions"
        ],
        "summary": "Get institution by MEC code",
        "parameters": [
          {
            "name": "codigoIes",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get institution by MEC code"
          }
        }
      }
    },
    "/api/v1/mec/institutions/{codigoIes}/courses": {
      "get": {
        "operationId": "mec-institutions_listCoursesByInstitutionCode",
        "tags": [
          "mec-institutions"
        ],
        "summary": "Get courses by institution",
        "parameters": [
          {
            "name": "codigoIes",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get courses by institution"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserFullPreferencesResponseDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "emailNotifications": {
            "type": "boolean"
          },
          "marketingEmails": {
            "type": "boolean"
          }
        }
      },
      "UserPreferencesResponseDto": {
        "type": "object",
        "properties": {
          "palette": {
            "type": "string"
          },
          "bannerColor": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "photoURL": {
            "type": "string"
          }
        }
      },
...
    }
  }
}